requirements.txt
awsebcli==3.21.0
blessed==1.20.0
blinker==1.9.0
botocore==1.35.99
cement==2.10.14
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.4
colorama==0.4.6
dlib==19.24.6
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==3.1.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==3.0.2
numpy==2.0.2
pathspec==0.10.1
pillow==11.1.0
pybind11==2.13.6
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
semantic-version==2.10.0
six==1.17.0
termcolor==2.5.0
urllib3==1.26.20
wcwidth==0.2.13
Werkzeug==3.1.3
zipp==3.21.0
Flask-Cors==5.0.0